dump & log collector bug fixes
- dump and log collector were still using the obsolete json_generator. replace that with the
dynamically generated json inventory file
- a few other fixes in dump
- new dump unittest

+++ kollacli/common.py
@@ -26,7 +26,6 @@ from kollacli.utils import get_kolla_etc
 from kollacli.utils import get_kolla_home
 from kollacli.utils import get_kolla_log_dir
 from kollacli.utils import get_kollacli_etc
-from kollacli.utils import get_kollacli_home
 from kollacli.utils import run_cmd
 
 from cliff.command import Command
@@ -122,11 +121,10 @@ class Dump(Command):
             kolla_logs = get_kolla_log_dir()
             kolla_ansible = os.path.join(kolla_home, 'ansible')
             kolla_docs = os.path.join(kolla_home, 'docs')
-            kolla_templates = os.path.join(kolla_home, 'templates')
             kolla_etc = get_kolla_etc()
             kolla_config = os.path.join(kolla_etc, 'config')
             kolla_globals = os.path.join(kolla_etc, 'globals.yml')
-            kollacli_etc = get_kollacli_etc()
+            kollacli_etc = get_kollacli_etc().rstrip('/')
             ketc = 'kolla/etc/'
             kshare = 'kolla/share/'
             fd, dump_path = tempfile.mkstemp(prefix='kollacli_dump_',
@@ -136,21 +134,18 @@ class Dump(Command):
                 # Can't blanket add kolla_home because the .ssh dir is
                 # accessible by the kolla user only (not kolla group)
                 tar.add(kolla_ansible,
-                        arcname=ketc + os.path.basename(kolla_ansible))
+                        arcname=kshare + os.path.basename(kolla_ansible))
                 tar.add(kolla_docs,
-                        arcname=ketc + os.path.basename(kolla_docs))
-                if os.path.isdir(kolla_templates):
-                    tar.add(kolla_templates,
-                            arcname=ketc + os.path.basename(kolla_templates))
+                        arcname=kshare + os.path.basename(kolla_docs))
 
                 # Can't blanket add kolla_etc because the passwords.yml
                 # file is accessible by the kolla user only (not kolla group)
                 tar.add(kolla_config,
-                        arcname=kshare + os.path.basename(kolla_config))
+                        arcname=ketc + os.path.basename(kolla_config))
                 tar.add(kolla_globals,
-                        arcname=kshare + os.path.basename(kolla_globals))
+                        arcname=ketc + os.path.basename(kolla_globals))
                 tar.add(kollacli_etc,
-                        arcname=kshare + os.path.basename(kollacli_etc))
+                        arcname=ketc + os.path.basename(kollacli_etc))
 
                 # add kolla log files
                 if os.path.isdir(kolla_logs):
@@ -174,22 +169,31 @@ class Dump(Command):
                 'kollacli password list']
 
         # collect the json inventory output
-        json_cmd = os.path.join(get_kollacli_home(), 'tools',
-                                'json_generator.py')
-        cmds.append(json_cmd)
-        fd, path = tempfile.mkstemp(suffix='.tmp')
-        os.close(fd)
-        with open(path, 'w') as tmp_file:
-            for cmd in cmds:
-                err_msg, output = run_cmd(cmd, False)
-                tmp_file.write('\n\n$ %s\n' % cmd)
-                if err_msg:
-                    tmp_file.write('Error message: %s\n' % err_msg)
-                for line in output:
-                    tmp_file.write(line + '\n')
+        inventory = Inventory.load()
+        inv_path = inventory.create_json_gen_file()
+        cmds.append(inv_path)
 
-        tar.add(path, arcname=os.path.join('kolla', 'cmds_output'))
-        os.remove(path)
+        try:
+            fd, path = tempfile.mkstemp(suffix='.tmp')
+            os.close(fd)
+            with open(path, 'w') as tmp_file:
+                for cmd in cmds:
+                    err_msg, output = run_cmd(cmd, False)
+                    tmp_file.write('\n\n$ %s\n' % cmd)
+                    if err_msg:
+                        tmp_file.write('Error message: %s\n' % err_msg)
+                    lines = output.split('\n')
+                    for line in lines:
+                        tmp_file.write(line + '\n')
+            tar.add(path, arcname=os.path.join('kolla', 'cmds_output'))
+
+        except Exception as e:
+            raise e
+        finally:
+            if path:
+                os.remove(path)
+            if inv_path:
+                os.remove(inv_path)
         return
